chore: refactor POS Quick Actions component
